Copy basic HTML codes from previous version and create the VB Script
For VB Script, the page gets the user name by cookies that are created from the previous page.
Also, request maxNum from the previous page, and create a cookie for that if the request is successful.
Moreover, create a random number based on the maxNum.
A random number is also created for a cookie. So,  firstly it checks whether the cookie already created or not