Move initializing conditions into the maxNum condtion
The cookies that were created in the previous game are still remaining. So, they need to be re-initialized.
That process should be done with the first submission, and it is the same time the maxNum=Request.Form("maxNum") is not a blank.
So, I moved all conditions in the same scope